requirements.txt
python-dotenv==1.1.0
colorama==0.4.6
PyMuPDF==1.26.0
python-docx==1.1.2
requests==2.32.3
bibtexparser==1.4.3
anthropic==0.52.1
openai==1.82.1
google-generativeai==0.8.5
click==8.2.1
tqdm==4.67.1
datasets==2.18.0
huggingface-hub==0.20.3
urllib3==2.2.1
pybliometrics==3.5.2
# WoS dependencies commented out as integration is currently inactive
# wos==0.4.1
requests-cache==1.1.0
tenacity==8.2.3
libcst==1.1.0
